requirements.txt
# Web Framework
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
colorama==0.4.6

# Cryptography
cryptography>=41.0.0
argon2-cffi==23.1.0

# GUI Framework
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.11.0

# Icons and UI
qtawesome>=1.2.0

# HTTP Requests
requests>=2.31.0